Refactoring code structure
