+++ packages/bff/src/api/tenantApiConverters.ts
@@ -1,18 +1,84 @@
+import { TenantWithOnlyAttributes } from "pagopa-interop-agreement-lifecycle";
 import {
-  tenantApi,
-  bffApi,
   attributeRegistryApi,
+  bffApi,
+  tenantApi,
 } from "pagopa-interop-api-clients";
 import { isDefined } from "pagopa-interop-commons";
-import { TenantMail } from "pagopa-interop-models";
-import { getLatestTenantContactEmail } from "../modelMappingUtils.js";
+import {
+  CertifiedTenantAttribute,
+  DeclaredTenantAttribute,
+  TenantAttribute,
+  TenantMail,
+  VerifiedTenantAttribute,
+  tenantAttributeType,
+  unsafeBrandId,
+} from "pagopa-interop-models";
+import { getLatestTenantContactEmail } from "../model/modelMappingUtils.js";
 
-export const toBffApiCompactOrganization = (
-  input: tenantApi.Tenant
-): bffApi.CompactOrganization => ({
-  id: input.id,
-  name: input.name,
-});
+export function toTenantAttribute(
+  att: tenantApi.TenantAttribute
+): TenantAttribute[] {
+  const certified: CertifiedTenantAttribute | undefined = att.certified && {
+    id: unsafeBrandId(att.certified.id),
+    type: tenantAttributeType.CERTIFIED,
+    revocationTimestamp: att.certified.revocationTimestamp
+      ? new Date(att.certified.revocationTimestamp)
+      : undefined,
+    assignmentTimestamp: new Date(att.certified.assignmentTimestamp),
+  };
+
+  const verified: VerifiedTenantAttribute | undefined = att.verified && {
+    id: unsafeBrandId(att.verified.id),
+    type: tenantAttributeType.VERIFIED,
+    assignmentTimestamp: new Date(att.verified.assignmentTimestamp),
+    verifiedBy: att.verified.verifiedBy.map((v) => ({
+      id: v.id,
+      verificationDate: new Date(v.verificationDate),
+      expirationDate: v.expirationDate ? new Date(v.expirationDate) : undefined,
+      extensionDate: v.extensionDate ? new Date(v.extensionDate) : undefined,
+    })),
+    revokedBy: att.verified.revokedBy.map((r) => ({
+      id: r.id,
+      verificationDate: new Date(r.verificationDate),
+      revocationDate: new Date(r.revocationDate),
+      expirationDate: r.expirationDate ? new Date(r.expirationDate) : undefined,
+      extensionDate: r.extensionDate ? new Date(r.extensionDate) : undefined,
+    })),
+  };
+
+  const declared: DeclaredTenantAttribute | undefined = att.declared && {
+    id: unsafeBrandId(att.declared.id),
+    type: tenantAttributeType.DECLARED,
+    assignmentTimestamp: new Date(att.declared.assignmentTimestamp),
+    revocationTimestamp: att.declared.revocationTimestamp
+      ? new Date(att.declared.revocationTimestamp)
+      : undefined,
+  };
+
+  return [certified, verified, declared].filter(
+    (a): a is TenantAttribute => !!a
+  );
+}
+
+export function toTenantWithOnlyAttributes(
+  tenant: tenantApi.Tenant
+): TenantWithOnlyAttributes {
+  return {
+    ...tenant,
+    attributes: tenant.attributes.map(toTenantAttribute).flat(),
+  };
+}
+
+export function toBffCompactOrganization(
+  tenant: tenantApi.Tenant
+): bffApi.CompactOrganization {
+  return {
+    id: tenant.id,
+    name: tenant.name,
+    kind: tenant.kind,
+  };
+}
 
 export const toBffApiRequesterCertifiedAttributes = (
   input: tenantApi.CertifiedAttribute
