+++ packages/bff/src/services/agreementService.ts
@@ -19,26 +19,30 @@ import {
 import {
   AgreementProcessClient,
   PagoPAInteropBeClients,
-} from "../providers/clientProvider.js";
+} from "../clients/clientsProvider.js";
 import { BffAppContext, Headers } from "../utilities/context.js";
-import { isAgreementUpgradable } from "../model/validators.js";
+
 import {
   agreementDescriptorNotFound,
   contractException,
   contractNotFound,
   invalidContentType,
-} from "../model/domain/errors.js";
-import {
-  toCompactDescriptor,
-  toCompactEservice,
-  toCompactEserviceLight,
-  toCompactOrganization,
-} from "../model/api/converters/catalogClientApiConverter.js";
+} from "../model/errors.js";
 import { config } from "../config/config.js";
 import { contentTypes } from "../utilities/mimeTypes.js";
 import { getLatestTenantContactEmail } from "../model/modelMappingUtils.js";
+
+import {
+  toCompactEservice,
+  toCompactDescriptor,
+} from "../api/catalogApiConverter.js";
+import {
+  toBffCompactOrganization,
+  toCompactEserviceLight,
+} from "../api/agreementApiConverter.js";
 import { getBulkAttributes } from "./attributeService.js";
 import { enhanceTenantAttributes } from "./tenantService.js";
+import { isAgreementUpgradable } from "./validators.js";
 
 export function agreementServiceBuilder(
   clients: PagoPAInteropBeClients,
@@ -458,7 +462,7 @@ export function agreementServiceBuilder(
           offset,
           totalCount: producers.totalCount,
         },
-        results: producers.results.map((p) => toCompactOrganization(p)),
+        results: producers.results.map((p) => toBffCompactOrganization(p)),
       };
     },
 
@@ -489,7 +493,7 @@ export function agreementServiceBuilder(
           offset,
           totalCount: consumers.totalCount,
         },
-        results: consumers.results.map((c) => toCompactOrganization(c)),
+        results: consumers.results.map((c) => toBffCompactOrganization(c)),
       };
     },
   };
