--- packages/bff/src/services/bffService.ts
@@ -1,5 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-export function bffServiceBuilder() {
-  return {};
-}
-export type BffService = ReturnType<typeof bffServiceBuilder>;
