+++ packages/bff/src/utilities/errorMappers.ts
@@ -2,7 +2,7 @@
 import { constants } from "http2";
 import { ApiError, CommonErrorCodes } from "pagopa-interop-models";
 import { match } from "ts-pattern";
-import { ErrorCodes as BFFErrorCodes } from "../model/domain/errors.js";
+import { ErrorCodes as BFFErrorCodes } from "../model/errors.js";
 
 type ErrorCodes = BFFErrorCodes | CommonErrorCodes;
