+++ packages/bff/src/utilities/fileUtils.ts
@@ -8,11 +8,12 @@ import AdmZip from "adm-zip";
 import { catalogApi } from "pagopa-interop-api-clients";
 import { FileManager, Logger } from "pagopa-interop-commons";
 import { DescriptorId, genericError } from "pagopa-interop-models";
-import { missingInterface } from "../model/domain/errors.js";
-import { verifyExportEligibility } from "../model/validators.js";
+import { missingInterface } from "../model/errors.js";
+import { verifyExportEligibility } from "../services/validators.js";
 import { retrieveEserviceDescriptor } from "../services/catalogService.js";
-import { ConfigurationEservice } from "../model/api/apiTypes.js";
-/* 
+import { ConfigurationEservice } from "../model/types.js";
+
+/*
   FileDocumentsRegistry is a map that contains the following information:
   - occurrences: a map that contains the number of occurrences of a document name
     (the same document name can be used multiple times in the same descriptor)
@@ -39,7 +40,7 @@ function getUniqueNameByDocumentId(
   return uniqueName;
 }
 
-/* 
+/*
   This function creates a FileDocumentsRegistry object that contains the following information:
   - occurrences: a map that contains the number of occurrences of a document name
     (the same document name can be used multiple times in the same descriptor)
@@ -118,8 +119,8 @@ export function buildJsonConfig(
   };
 }
 
-/* 
-  This function creates a zip file fetched from the S3 bucket 
+/*
+  This function creates a zip file fetched from the S3 bucket
   using FileManager, the zip file containing the following files:
   - descriptor's interface file
   - descriptor's documents file
